# GoogleFormsApi/openapi.yml
openapi: 3.0.1
info:
  title: 'GoogleFormsApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null'
  version: '1.0'
paths:
  /api/Account/IsLoggedInn:
    get:
      tags:
        - Account
      parameters:
        - name: name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Success
...
  /api/Account/Register:
    post:
      tags:
        - Account
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
      responses:
        '200':
          description: Success
  /api/Account/Login:
    post:
      tags:
        - Account
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Success
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
    RegistrationRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        confirmPassword:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
      additionalProperties: false
...